% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amstar2_barplot.R
\name{amstar2_barplot}
\alias{amstar2_barplot}
\title{Plots the AMSTAR plot}
\usage{
amstar2_barplot(
  data = data,
  fontsize = 3.5,
  fontcolor = "white",
  barwidth = 0.8
)
}
\arguments{
\item{data}{A dataset.}

\item{fontsize}{A number that controls the aesthetic of font size of the percentages in the bars. Default is fontsize = 3.5 .}

\item{fontcolor}{A color that controls the aesthetic of color of the percentages in the bars. Default is fontcolor = "white".}

\item{barwidth}{A number that controls the width between the bars. Default is barwidth = 0.8 .}
}
\value{
amstar2_barplot
}
\description{
This function plots the AMSTAR bar plot.
}
