% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amstar2_donut.R
\name{amstar2_donut}
\alias{amstar2_donut}
\title{Plots the AMSTAR donut plot.}
\usage{
amstar2_donut(data = data, r1 = 1, r2 = 0.75)
}
\arguments{
\item{data}{A dataset.}

\item{r1}{Outer circle radius.}

\item{r2}{Inner circle radius, should inferior  to `r1` value.}
}
\value{
donut
}
\description{
Plots the AMSTAR donut plot.
}
