% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amstar2_gtable.R
\name{amstar2_gtable}
\alias{amstar2_gtable}
\title{Generated a AMSTAR 2 gt table}
\usage{
amstar2_gtable(data = data)
}
\arguments{
\item{data}{A dataset.}
}
\value{
amstar_gtable
}
\description{
This function creates a AMSTAR 2 gt table.
}
\examples{
DATASET <- readxl::read_excel(system.file("extdata", "dat.xlsx", package = "amstar2"))

amstar2_gtable(DATASET)
}
