% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amstar2_halfdonut.R
\name{amstar2_halfdonut}
\alias{amstar2_halfdonut}
\title{Plots the AMSTAR halfdonut plot.}
\usage{
amstar2_halfdonut(data = data, r1 = 1, r2 = 0.75)
}
\arguments{
\item{data}{A dataset.}

\item{r1}{Outer circle radius.}

\item{r2}{Inner circle radius, should inferior  to `r1` value.}
}
\value{
halfdonut
}
\description{
Plots the AMSTAR halfdonut plot.
}
