% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amstar2_overall.R
\name{amstar2_overall}
\alias{amstar2_overall}
\title{Plots the AMSTAR overall plot.}
\usage{
amstar2_overall(
  data = data,
  r1 = 1,
  r2 = 0.75,
  table_size = 14,
  caption_size = 12
)
}
\arguments{
\item{data}{A dataset.}

\item{r1}{Outer circle radius. Default is `r1 = 1`.}

\item{r2}{Inner circle radius, should inferior  to `r1` value. Default is `r2 = 0.75`.}

\item{table_size}{It controls the size of the inset table. Default is `table_size = 14`.}

\item{caption_size}{It controls the font size of the caption. Default is `caption_size = 12`.}
}
\value{
halfdonut
}
\description{
Plots the AMSTAR overall plot.
}
\examples{
DATASET <- readxl::read_excel(system.file("extdata", "dat.xlsx", package = "amstar2"))

amstar2_overall(DATASET)
}
