% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amstar2_plot.R
\name{amstar2_plot}
\alias{amstar2_plot}
\title{Plots the AMSTAR plot}
\usage{
amstar2_plot(data = data, fontsize = 3.5)
}
\arguments{
\item{data}{A dataset.}

\item{fontsize}{A number which controls the aesthetic of font size of the percentages in the bars.}
}
\value{
amstar2_plot
}
\description{
This function plots the AMSTAR bar plot.
}
