% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amstar2_table.R
\name{amstar2_table}
\alias{amstar2_table}
\title{Calculates the overall rating based on critical items}
\usage{
amstar2_table(data = data)
}
\arguments{
\item{data}{A dataset.}
}
\value{
amstar_table
}
\description{
This function adds the overall assessment.
}
\examples{
DATASET <- readxl::read_excel(system.file("extdata", "dat.xlsx", package = "amstar2"))

tb <- amstar2_table(DATASET)
}
