% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amstar2_gtable.R
\name{amstar2_gtable}
\alias{amstar2_gtable}
\title{Generates the AMSTAR 2 main table}
\usage{
amstar2_gtable(data = data)
}
\arguments{
\item{data}{A dataset containing the item ratings.}
}
\value{
A “gt” table object.
}
\description{
It creates a nice-looking “gt” table that includes the item ratings and the overall confidence rating for each SR according to AMSTAR 2 guidelines.
}
\examples{
library(amstar2)

DATASET <- readxl::read_excel(system.file("extdata", "sample_dat.xlsx", package = "amstar2"))

amstar2_gtable(DATASET)
}
