% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amstar2_plot.R
\name{amstar2_plot}
\alias{amstar2_plot}
\title{Plots the AMSTAR plot}
\usage{
amstar2_plot(data = data)
}
\arguments{
\item{data}{A dataset.}
}
\value{
amstar2_plot
}
\description{
This function plots the AMSTAR bar plot.
}
